+++ inc/ti/fn/fntypecount.h
@@ -16,7 +16,21 @@ static int do__f_type_count(ti_query_t * query, cleri_node_t * nd, ex_t * e)
     if (!type)
         return ti_raw_err_not_found((ti_raw_t *) query->rval, "type", e);
 
-    n = ti_query_count_type(query, type);
+    /* use cache if available */
+    if (type->t_cache)
+    {
+        n = type->t_cache->n;
+    }
+    else
+    {
+        ssize_t r = ti_query_count_type(query, type);
+        if (r < 0)
+        {
+            ex_set_mem(e);
+            return e->nr;
+        }
+        n = (size_t) r;
+    }
 
     ti_val_unsafe_drop(query->rval);
     query->rval = (ti_val_t *) ti_vint_create((int64_t) n);