+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # v1.8.6-alpha0
 
-* Read `type_count(..)` from cache if available.
+* Read `type_count(..)` from cache if available, pr #443.
 
 # v1.8.5